requirements.txt
# Core dependencies
streamlit>=1.31.0
llama-index-core>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
python-dotenv>=1.0.0

# Document parsing
llama-parse>=0.3.0
pypdf>=4.0.0
python-docx>=1.0.0
openpyxl>=3.1.2
python-pptx>=0.6.21
pandas>=2.2.0

# Code highlighting
Pygments>=2.17.0

# Machine learning
torch>=2.2.0
transformers>=4.37.0
sentence-transformers>=2.5.0

# Email parsing
mbox-reader>=0.1.0
email-parser>=0.1.3

# Image processing
Pillow>=10.0.0

# Additional formats
python-magic>=0.4.27
chardet>=5.2.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
PyYAML>=6.0.1

# Archive handling
py7zr>=0.20.0
rarfile>=4.0

# Web framework
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# URL processing
requests>=2.31.0
validators>=0.22.0